Download Resource functionality

+++ src/main/java/es/hugoalvarezajenjo/selecta/services/resources/repository/SubjectResourceRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SubjectResourceRepository extends JpaRepository<SubjectResource, Integer> {
+public interface SubjectResourceRepository extends JpaRepository<SubjectResource, Long> {
     List<SubjectResource> findSubjectResourceBySubjectId(Long subjectId);
 }
